pyproject.toml
[project]
name = "butadon-backend-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "chromadb>=1.0.15",
    "fastapi>=0.116.1",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.0.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "fastapi-cli>=0.0.8",
    "cuid>=0.4.0",
    "openai>=1.97.0",
    "python-multipart>=0.0.20",
    "pypdf2>=3.0.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "pycryptodome>=3.23.0",
    "tiktoken>=0.9.0",
    "pypdf>=5.8.0",
]
